CXP-1427: move components into a new folder and fix a counter
