+++ src/Akeneo/Connectivity/Connection/front/src/connect/components/CustomApps/CustomAppCard.tsx
