+++ src/Akeneo/Connectivity/Connection/front/src/connect/components/CustomApps/CustomAppList.tsx
