+++ src/Akeneo/Connectivity/Connection/front/tests/src/connect/components/CustomApp/CustomAppCard.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import {screen} from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
 import {renderWithProviders} from '../../../../test-utils';
-import {CustomAppCard} from '@src/connect/components/CustomApp/CustomAppCard';
+import {CustomAppCard} from '@src/connect/components/CustomApps/CustomAppCard';
 import {SecurityContext} from '@src/shared/security';
 
 test('it displays test app', () => {
